Remove code duplication in dist.mc
Removes excessive code duplication and change the `Dist` fragment so that it is
easier to extend and maintain it. Previously each concrete distribution was
essentially treated as a uniqe term.

Type-checking of distributions is now similar to how we type check constants in
MCore. In addition this change fixes a bug where parameters of type checked
distributions would not include type annotations.