Remove excessive code duplication in dist.mc and add type information for inference methods  (#183)
* Add builder functions to elementary functions fragment

* Add inference method record type function

* Remove code duplication in dist.mc

Removes excessive code duplication and change the `Dist` fragment so that it is
easier to extend and maintain it. Previously each concrete distribution was
essentially treated as a uniqe term.

Type-checking of distributions is now similar to how we type check constants in
MCore. In addition this change fixes a bug where parameters of type checked
distributions would not include type annotations.

* Correct names for shallow map/folds